test(api): [API-022] :truck: Move controllers unit tests in test/uni/controllers
